+++ AgroTemp.Application/Dtos/ProbeWithDetailsDto.cs
@@ -4,6 +4,6 @@ public class ProbeWithDetailsDto : ProbeDto
 {
     public SiloDto Silo { get; set; }
     public ReadingModuleDto ReadingModule { get; set; }
-    public List<ushort?> ActualTemperatures { get; set; }
-    public List<ushort?> ActualDeltaTemperatures { get; set; }
+    public List<double?> ActualTemperatures { get; set; }
+    public List<double?> ActualDeltaTemperatures { get; set; }
 }
